adds zobrist keys and some basic internal move ordering
transposition table left on another branch. it will be redone. other things need to be implemented first and code needs to be cleaned up.

internal move ordering is move ordering that happens in the scope of the move generation. it currently interferes with perft tests but increases search speed singificantly.